+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 ### Features
 - cta/CTA#1205 - Fail pipeline if cppcheck detects errors
 - cta/CTA#1206 - Change NULL for nullptr
+- cta/CTA#1217 - Schema verification should just issue an warning if there are extra indexes in the db that are not in the catalogue
 ### Bug fixes
 ### Building and Packaging
 ### Catalogue Schema